backend/requirements.txt
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
fastapi-mail==1.6.2
greenlet==3.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pyotp==2.9.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
regex==2025.11.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.46
standard-imghdr==3.13.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
wrapt==2.1.1
